requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-socketio[asgi]==5.11.4
jinja2==3.1.4
